Move statistics related functions to analytics.c
