+++ daemon/analytics.h
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef NETDATA_ANALYTICS_H
+#define NETDATA_ANALYTICS_H 1
+
+#include "../daemon/common.h"
+
+extern void set_global_environment();
+extern void send_statistics( const char *action, const char *action_result, const char *action_data);
+
+#endif //NETDATA_ANALYTICS_H
