+++ daemon/common.h
@@ -82,6 +82,7 @@
 #include "main.h"
 #include "signals.h"
 #include "commands.h"
+#include "analytics.h"
 
 // global netdata daemon variables
 extern char *netdata_configured_hostname;
