+++ libnetdata/log/log.h
@@ -40,6 +40,7 @@ extern "C" {
 #define D_STATSD            0x0000000010000000
 #define D_POLLFD            0x0000000020000000
 #define D_STREAM            0x0000000040000000
+#define D_ANALYTICS         0x0000000080000000
 #define D_RRDENGINE         0x0000000100000000
 #define D_ACLK              0x0000000200000000
 #define D_METADATALOG       0x0000000400000000
